feat: add lymph model & form to get it from repo
It is now possible to specify a git repo URL along with a revision where
inferred samples are stored in an HDF5 file and all relevant model
parameters are kept in a YAML file. The Django model created from this
will then compute the prior state probabilities under that model and use
it for risk predictions in the dashboard.

Related: #15